feat: move foundry $ref resolver and edit helper to pkg/foundry
